ServiceB/requirements.txt
fastapi
uvicorn
httpx
dapr
dapr-ext-grpc